Added OptiFBS option for stochastic (or ''partial'') gradient computation from a subset of mapsCell in CostSummation.
Example: CostSummation of 61 Cost with the 61st mapsCell being a CostHyperbolic and the others are L2 Cost. The stochastic gradient scheme can be applied on the L2 Costs only (1 to 60) by setting the set attribute to [1:60]. The gradient at each iteration will be from a subset of these 60 L2 Costs and the CostHyperbolic.