Add 2D X-ray operator based on MATLAB's radon and iradon
See the test for a small example of how to use it. Note
that the adjoint is not really the adjoint, but only an
approximation (up to about 20 dB). If we want something cleaner
we can use astra-toolbox, but this would require the user to
install something (and possibly compile it from source).